a lot, everywhere....
User signing in will also save the id of said user in the browser
this is so that the user is saved in the poll, as the owner. which will be needed next
this needed changes all across the board so there is bs everywhere haha

also added a random key generator, so if the user doesn't sign in but wants to vote I got 2 keys to recognize him by. this is probably completely unnecessary but I realized that too late myself so here we are.

oh, a not signed in voter will now have his ip and a random number saved in a cookie, this is saved in the db and in return the id of said user is saved by the user, this is used for voting. This will need some polishing but it is working for now.

in a later patch this will also be linked to a user on signup. said user can then signout but the voterid won't be disposed of, to make sure the votes are fair.